+++ NEWS.md
@@ -1,5 +1,6 @@
 # AzureAuth 1.0.2.9000
 
+* Support providing a certificate via a PEM file, or an AzureKeyVault object.
 * Support providing a path in the aad_host argument, for B2C logins.
 
 # AzureAuth 1.0.2